pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hedge-fund-ml"
version = "0.1.0"
description = "Research utilities for hedge fund machine learning experiments"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [{name = "Research Team"}]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "matplotlib>=3.7",
  "pydantic>=2.4",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "black>=23.12",
  "mypy>=1.7",
  "pip-tools>=7.3",
  "pre-commit>=3.6",
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.1.8",
]
deep-learning = [
  "tensorflow>=2.13",
  "keras>=2.13",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["hedge_fund_ml*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
target-version = "py310"
line-length = 88
exclude = ["GAN"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_equality = true
explicit_package_bases = true
mypy_path = ["src"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "matplotlib.*",
  "keras.*",
  "tensorflow.*",
  "yaml",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
  "ignore::DeprecationWarning",
]
